Clean up before finishing framebuffer stuff.

+++ InariKonKon/include/InariKonKon/Utility/Time.hpp
@@ -26,15 +26,15 @@ namespace ikk
 		constexpr ~Time() noexcept = default;
 
 		inline constexpr float asSeconds() const noexcept				{ return static_cast<float>(this->m_value.count() / 1000000.f);		};
-		inline constexpr std::uint32_t asMilliseconds() const noexcept	{ return static_cast<std::uint32_t>(this->m_value.count() / 1000u);	};
-		inline constexpr std::uint64_t asMicroseconds() const noexcept	{ return static_cast<std::uint64_t>(this->m_value.count());			};
+		inline constexpr std::int32_t asMilliseconds() const noexcept	{ return static_cast<std::int32_t>(this->m_value.count() / 1000u);	};
+		inline constexpr std::int64_t asMicroseconds() const noexcept	{ return this->m_value.count();										};
 
 		inline constexpr std::chrono::microseconds toDuration() const noexcept { return this->m_value; };
 	private:
 		std::chrono::microseconds m_value{};
 	};
 
-	[[nodiscard]] inline constexpr Time seconds(const float amount) noexcept				{ return Time(amount); };
-	[[nodiscard]] inline constexpr Time milliseconds(const std::uint32_t amount) noexcept	{ return Time(amount); };
-	[[nodiscard]] inline constexpr Time microseconds(const std::uint64_t amount) noexcept	{ return Time(amount); };
+	[[nodiscard]] inline constexpr Time seconds(const float amount)				 noexcept { return Time(amount); };
+	[[nodiscard]] inline constexpr Time milliseconds(const std::uint32_t amount) noexcept { return Time(amount); };
+	[[nodiscard]] inline constexpr Time microseconds(const std::uint64_t amount) noexcept { return Time(amount); };
 }
